requirements.txt
anthropic==0.8.1
pydantic==2.7.4
pydantic-core==2.18.4
pyyaml==6.0.1
tenacity==8.2.3
cohere==5.13.11
pandas==2.2.1
python-dotenv==1.0.1 